homework2/movie_theater_booking/requirements.txt
asgiref==3.7.2
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
dj-database-url==3.0.1
Django==4.2.11
djangorestframework==3.16.1
gunicorn==23.0.0
gyp==0.1
h11==0.16.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pytest==8.4.2
pytz==2024.1
requests==2.32.5
setuptools==68.1.2
six==1.16.0
sqlparse==0.4.4
supervisor==4.2.5
urllib3==2.5.0
uvicorn==0.37.0
wheel==0.42.0
whitenoise==6.11.0
